- image: added helper functions to image to get datetime and geolocation information with ease - image tags: added offsets in front of tag ids to namespace them to avoid conflicts. this is a wip - core datetime: fixed issue with ISO 1806 strings if no - was present - freeimage plugin: fixed metadata array parsing of multiple values and added tag id mapping functions - FileSorter: Now uses image property datetime if available to potentially override the filesystem datetime

+++ tools/udp2rediscluster/src/udp2rediscluster.cpp
@@ -2023,7 +2023,7 @@ Udp2RedisCluster::LoadConfig( const CORE::CDataNode& globalConfig )
     m_redisStreamStartChannelID = appConfig->GetAttributeValueOrChildValueByName( "redisStreamStartChannelID" ).AsInt32( m_redisStreamStartChannelID, true );
 
     CORE::CString::StringSet channelIDStrs = appConfig->GetAttributeValueOrChildValueByName( "channelIDs" ).AsString( CORE::CString::Empty, true ).ParseUniqueElements( ',', false );
-    m_channelCount = appConfig->GetAttributeValueOrChildValueByName( "channelCount" ).AsUInt16( channelIDStrs.empty() ? 1 : channelIDStrs.size(), true );
+    m_channelCount = appConfig->GetAttributeValueOrChildValueByName( "channelCount" ).AsUInt16( (UInt16) channelIDStrs.empty() ? 1 : (UInt16) channelIDStrs.size(), true );
 
     m_udpStartPort = appConfig->GetAttributeValueOrChildValueByName( "udpStartPort" ).AsUInt16( m_udpStartPort, true );
     CORE::Int32 udpPortChannelIdOffset = appConfig->GetAttributeValueOrChildValueByName( "udpPortChannelIdOffset" ).AsInt32( 20000, true );
